The big reorg...continued...update to Aspose licensing
Hold lic file as a b64 encoded String in the LicenseHelper and make a temp file from it to create the licences. Update Version util to cope with ocr version. Add a log4j so we can see some debugs in testing? Bump ocr to 20.11 and diagram to 20.12 (fixes regression)