Beans updated for latest Aspose libs and WorkFolderManager fix
Added/removed a couple of classifiers (for HTML and CAD) in the pom (contrary to the information on Aspose web site). Added a couple of throws in DocumentBuilderUtils and a try-catch in TocIterator for some new Exceptions being thrown by Aspose writeln and getField. Corrected (finally, hopefully) the logic to deal with Work and Error folder validation, auto-creating them if needed and writing/deleting a test file.