Fix up the licence helper test, so it actually uses the (pdf) licence
Added helpful comment on getVersionFromPom about the fact you need to have done a build of beans for it to work. Changed the code around in LicenseHelper to do a try-with-resources on the VFS and to only call FileUtils delete if we used a temp file. Move setLocales and setLicences to separate methods. Added a TODO about the CurrentThreadSettings setLocale being deprecated.